# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CaribeTour API",
        "description": "API documentation for CaribeTour project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/types": {
            "get": {
                "summary": "Get list of types",
                "description": "Display a listing of the resource.",
                "operationId": "294c29e77d2424d2754db38dac833443",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter types by fields (e.g., name)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit the number of types returned",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Create a new type",
                "description": "Store a newly created resource in storage.",
                "operationId": "501ca968763db6d4e180c9c03e128b89",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Type Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Type created successfully"
                    }
                }
            }
        },
        "/api/v1/types/{id}": {
            "get": {
                "summary": "Get a type by ID",
                "description": "Display the specified resource.",
                "operationId": "444683c95a9bddb4bad4747e36991524",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update a type",
                "description": "Update the specified resource in storage.",
                "operationId": "ea224d615f6726887e9c2b31a736dc4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Type Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Type updated successfully"
                    }
                }
            },
            "delete": {
                "summary": "Delete a type",
                "description": "Remove the specified resource from storage.",
                "operationId": "a1303e5639393a10ed4a9319bc1ffe56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Type deleted successfully"
                    }
                }
            }
        }
    }
}